Refactor Dockerfile and devcontainer.json
This commit refactors the Dockerfile and devcontainer.json files. It updates the base image to mcr.microsoft.com/devcontainers/base:ubuntu-24.04 and adds a volume mount for docker-in-docker. Additionally, it modifies the entrypoint and command in the Dockerfile. In the devcontainer.json file, it changes the postCreateCommand to use sudo and fixes a typo in the mounts section. The commit also updates the post_create.sh script to use sudo for certain commands.